home_budget/requirements/base.txt
django-recaptcha==2.0.6
openpyxl==3.0.3
pygal==2.4.0
asgiref==3.2.3
Django==3.0.3
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
python-dateutil==2.8.1
celery
djangorestframework